+++ cpgloader-tinkergraph-shiftleft/src/main/scala/io/shiftleft/cpgloading/tinkergraphshiftleft/CpgLoader.scala
@@ -1,11 +1,11 @@
 package io.shiftleft.cpgloading.tinkergraphshiftleft
 
-import io.shiftleft.cpgloading.{CpgLoaderBase, ProtoToTinkerShiftleftCpg}
+import io.shiftleft.cpgloading.CpgLoaderBase
 
 /* singleton instance for convenience */
 object CpgLoader extends CpgLoader
 
 /** Load cpg proto (typically cpg.bin.zip) into a new Tinkergraph instance */
 class CpgLoader extends CpgLoaderBase {
-  override def builder = new ProtoToTinkerShiftleftCpg
+  override def builder = new ProtoToCpg
 }
