+++ cpgloader-tinkergraph/src/main/scala/io/shiftleft/cpgloading/tinkergraph/CpgLoader.scala
@@ -1,11 +1,11 @@
 package io.shiftleft.cpgloading.tinkergraph
 
-import io.shiftleft.cpgloading.{CpgLoaderBase, ProtoToTinkerCpg}
+import io.shiftleft.cpgloading.CpgLoaderBase
 
 /* singleton instance for convenience */
 object CpgLoader extends CpgLoader
 
 /** Load cpg proto (typically cpg.bin.zip) into a new Tinkergraph instance */
 class CpgLoader extends CpgLoaderBase {
-  override def builder = new ProtoToTinkerCpg
+  override def builder = new ProtoToCpg
 }
