+++ enhancements/src/main/scala/io/shiftleft/cpgloading/CpgLoaderBase.scala
@@ -7,7 +7,7 @@ import io.shiftleft.queryprimitives.steps.starters.Cpg
 /** Load cpg proto (typically cpg.bin.zip) into a graph instance */
 abstract class CpgLoaderBase {
 
-  def builder: ProtoToCpg
+  def builder: ProtoToCpgBase
 
   def loadCodePropertyGraph(filename: String, runEnhancements: Boolean = true): Cpg = {
     val cpg = new ProtoCpgLoader(builder).loadFromProtoZip(filename)
